Outlined a world map gameobject
The world map object maps a tileset to world coordinates. The user can specify how large they want a tile to be represented in world space. When the player enters the area of a new tile, the map updates the color of the tile to show it has been visited. It has also been implemented to function with the json save system to save and load this data. Still need to hook up the map to the main UI component